api methods are completed.

+++ task-manager/src/routers/task.js
@@ -2,9 +2,13 @@ const express = require("express");
 const Task = require("../models/task");
 const router = new express.Router();
 const checkForValidUpdates = require("../common-files/common-functions");
+const auth = require('../middleware/auth');
 
-router.post("/tasks", async (req, res) => {
-  const task = new Task(req.body);
+router.post("/tasks", auth, async (req, res) => {
+  const task = new Task({
+    ...req.body,
+    owner: req.user._id
+  });
   try {
     await task.save();
     res.status(201).send(task);
@@ -13,31 +17,34 @@ router.post("/tasks", async (req, res) => {
   }
 });
 
-router.get("/tasks", async (req, res) => {
+router.get("/tasks", auth, async (req, res) => {
   try {
-    const tasks = await Task.find({});
-    res.send(tasks);
+    await req.user.populate('tasks').execPopulate();
+    res.send(req.user.tasks);
   } catch (error) {
     res.status(500).send();
   }
 });
 
-router.get("/task/:id", async (req, res) => {
+router.get("/task/:id", auth, async (req, res) => {
   const _id = req.params.id;
   try {
-    const task = await Task.findById(_id);
+    const task = await Task.findOne({ _id, owner: req.user._id });
+    if (!task) {
+      res.status(404).send();
+    }
     res.send(task);
   } catch (error) {
     res.status(500).send();
   }
 });
 
-router.patch("/tasks/:id", async (req, res) => {
+router.patch("/tasks/:id", auth, async (req, res) => {
   if (!checkForValidUpdates(Task, req.body)) {
     return res.status(400).send({ error: "Invalid updates!" });
   }
   try {
-    const task = await Task.findById(_id);
+    const task = await Task.findById({ _id: req.params.id, owner: req.user._id });
     if (!task) {
       return res.status(404).send();
     }
@@ -51,10 +58,9 @@ router.patch("/tasks/:id", async (req, res) => {
   }
 });
 
-router.delete("/tasks/:id", async (req, res) => {
+router.delete("/tasks/:id", auth, async (req, res) => {
   try {
-    const task = await Task.findOneAndDelete({ _id: req.params.id });
-
+    const task = await Task.findOneAndDelete({ _id: req.params.id, owner: req.user._id});
     if (!task) {
       return res.status(404).send();
     }
